// frontend/gestor_material/src/app/models/produto.model.ts
export class ProdutoModel {
  id: number;
  idColigada: number;
  codigo: string;
  ean: string;
  descricao: string;
  descricaoTexto: string;
  unidade: string;
  idGrupo: number;
  idSubGrupo: number;
  apresentacao: string;
  informacaotecnica: string;
  pesoLiquido: number;
  pesoBruto: number;
  largura: number;
  altura: number;
  profundidade: number;
  estoque: number;
  precoVarejo: number;
  precoAtacado: number;
  fatorVenda: number;
  ativo: string;
  unidadeFator: string;
  promocao: string;
  imagem: string;
}